reduce number of files
